chore: move more tests to new setup
