+++ cmake/Native.cmake
@@ -13,10 +13,10 @@ target_link_libraries(${PROJECT_NAME}
     SDL2_mixer::SDL2_mixer-static)
 
 # Create test executable
-add_executable(test.out test/main.cpp ${TEST_LIB_FILES})
-target_link_libraries(test.out PRIVATE ${PROJECT_NAME})
-add_custom_target(test
-    COMMAND LSAN_OPTIONS=suppressions=asan.supp $<TARGET_FILE:test.out>
-    DEPENDS test
-    WORKING_DIRECTORY ${CMAKE_SOURCE_DIR}
-)
+# add_executable(test.out test/main.cpp ${TEST_LIB_FILES})
+# target_link_libraries(test.out PRIVATE ${PROJECT_NAME})
+# add_custom_target(test
+#     COMMAND LSAN_OPTIONS=suppressions=asan.supp $<TARGET_FILE:test.out>
+#     DEPENDS test
+#     WORKING_DIRECTORY ${CMAKE_SOURCE_DIR}
+# )
