+++ src/core/geometry/Box.test.cpp
@@ -1,9 +1,9 @@
 #include "core/geometry/Box.h"
 
-#include "../lib/test.h"
+#include "../../../test/lib/test.h"
 #include "core/geometry/Vector.h"
 
-void constructor_test() {
+void constructor() {
   Box box;
   assertEq("default constructor sets top left corner to (0, 0)",
            box.getCorner(), Vector());
@@ -11,11 +11,13 @@ void constructor_test() {
   assertEq("default constructor sets width to 0", box.getWidth(), 0.0);
 
   assertEq("default constructor sets height to 0", box.getHeight(), 0.0);
+}
 
+void parametrizedConstructor() {
   Vector toleft(1, 2);
   float width = 3.5;
   float height = 4.5;
-  box = Box(toleft, width, height);
+  Box box = Box(toleft, width, height);
 
   assertEq("parameterized constructor sets top left corner correctly",
            box.getCorner(), toleft);
@@ -25,30 +27,32 @@ void constructor_test() {
            height);
 }
 
-void setCorner_test() {
+void setCorner() {
   Box box;
   Vector corner(5, 6);
   box.setCorner(corner);
   assertEq("sets top left corner correctly", box.getCorner(), corner);
 }
 
-void setWidth_test() {
+void setWidth() {
   Box box;
   float width = 10.5;
   box.setWidth(width);
   assertEq("sets width correctly", box.getWidth(), width);
 }
 
-void setHeight_test() {
+void setHeight() {
   Box box;
   float height = 15.5;
   box.setHeight(height);
   assertEq("sets height correctly", box.getHeight(), height);
 }
 
-void Box_test() {
-  test("constructor", constructor_test);
-  test("setCorner", setCorner_test);
-  test("setWidth", setWidth_test);
-  test("setHeight", setHeight_test);
+auto main() -> int {
+  test("constructor", constructor);
+  test("parametrized constructor", parametrizedConstructor);
+  test("setCorner", setCorner);
+  test("setWidth", setWidth);
+  test("setHeight", setHeight);
+  return report();
 }
