+++ test/suites/ObjectList_test.cpp
@@ -1,6 +1,6 @@
 #include "core/objects/ObjectList.h"
 
-#include "../lib/test.h"
+#include "../../../test/lib/test.h"
 
 using namespace lb;
